pyproject.toml
[project]
name = "krasip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20240930",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "sh>=2.0.7",
]
